Fixed Employee controller id routing bug. Edit,delete,view and add employee added

+++ Models/ProductModels/Product.cs
@@ -10,7 +10,7 @@ namespace Project_C.Models.ProductModels
         public Guid Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public Decimal Price { get; set; }
+        public double Price { get; set; }
         public string Place { get; set; }
         public string? PhotoPath { get; set; }
         public string? VideoLink { get; set; }
